+++ packaging/settings.py
@@ -336,6 +336,22 @@ if config.getboolean('logging', 'syslog'):
     LOGGING['loggers']['django']['handlers'].append('syslog')
     LOGGING['loggers']['nodeconductor']['handlers'].append('syslog')
 
+if config.get('logging', 'log_level').upper() == 'DEBUG':
+    # Enabling debugging at http.client level (requests->urllib3->http.client)
+    # you will see the REQUEST, including HEADERS and DATA, and RESPONSE with HEADERS but without DATA.
+    # the only thing missing will be the response.body which is not logged.
+    try: # for Python 3
+        from http.client import HTTPConnection
+    except ImportError:
+        from httplib import HTTPConnection
+    HTTPConnection.debuglevel = 1
+
+    LOGGING['loggers']['requests.packages.urllib3'] = {
+        'handlers': ['file'],
+        'level': 'DEBUG',
+        'propagate': True
+    }
+
 if config.get('events', 'log_file') != '':
     LOGGING['handlers']['file-event']['filename'] = config.get('events', 'log_file')
     LOGGING['loggers']['nodeconductor']['handlers'].append('file-event')